Added ability to use not all principal components
There is 199 dimensions by default, but not all principal components are
needed during optimization and it costs a lot to concatenate array of
zeros to existent array with coefficients.

It was decided to suppose, that all unset coefficients are set to zero
and not iterate them during Face calculation in C functions.